+++ .gitignore
@@ -1,8 +1,11 @@
+.vscode/
+src/deployment_steps
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
 *$py.class
-.vscode/
+
 
 # C extensions
 *.so
