+++ requirements.txt
@@ -1,5 +1,5 @@
-snowflake-connector-python=3.10.1
-faker=25.5.0
-pandas=2.2.2
-dot-env=1.0.1
-boto3=1.34.119
+snowflake-connector-python==3.10.1
+faker==25.5.0
+pandas==2.2.2
+boto3==1.34.119
+python-dotenv==1.0.1
